key_server/pyproject.toml
[tool.poetry]
name = "key-server"
version = "0.1.0"
description = ""
authors = ["takitsu21 <dylannbatisse@gmail.com>"]
readme = "README.md"
packages = [{include = "key_server"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.95.1"
uvicorn = "^0.21.1"
sqlalchemy = "^2.0.10"
psycopg2-binary = "^2.9.6"
pycryptodome = "^3.17"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
